requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12,<3.13
drf-yasg>=1.20,<1.21
pandas>=1.2,<1.3